Kurek maven

+++ src/main/java/Model/Edge.java
@@ -2,17 +2,17 @@ package Model;
 
 public class Edge {
 
-    private int cost;
+    private double cost;
     private int firstVertexId;
     private int secondVertexId;
 
-    public Edge(int cost, int firstVertexId, int secondVertexId) {
+    public Edge(double cost, int firstVertexId, int secondVertexId) {
         this.cost = cost;
         this.firstVertexId = firstVertexId;
         this.secondVertexId = secondVertexId;
     }
 
-    public int getCost() {
+    public double getCost() {
         return this.cost;
     }
